FIX: instead of pip, using pypi for version check
There are issues with `pip search`, so now I am using `requests` module
instead to send a HTTP GET request to pypi website and pattern match
that to our versioning to get latest release version. Maybe there exists
a better way to do this, but until then, this is great.