Experimental: sample_config.toml in release
Currently I store default sample config as a string in constants.py.
Instead, I wanted to have a sample_config.toml file which should be read
at runtime, and the contents be written in the target config.toml file
in the project's home directory.

It's kinda difficult to test this, because I don't exactly get yet wth
is going on, but I hope this works :)