Added sample_config.toml to package
Took more time than it should, but issoke.

So, from now on, instead of a string in constants.py, we have a
sample_config.toml file that we can open for reading, and write the read
text to the desired location.

I don't know if there exists a better way to do all this, but this is
definitely good, at least for now.