Exam -> ETA and eta command added
More generalized functionality for the countdown. Also allows people to store the note,
and retrieve it by using command `eta` (`e` for shorthand).
People can set it by editing the config file. It'd be great if it was
editable directly from the CLI, instead of editing the config file
manually. I'll work on it.